refactor(api/files): extract dualAuth into shared middleware module (#4053)
Reviewers (principles, uncle-bob) flagged the dualAuth implementation as
40 lines of doc comment + runtime assertion living inline in the route
file — both because the pattern (API-key auth with session-cookie
fallback) will recur for any browser-served byte endpoint, and because
inlining hides the authMiddleware contract fragility behind a wall of
warnings every reader has to scroll past.

Moved to langwatch/src/app/api/middleware/dual-auth.ts.
No runtime behavior change. The credential-less-skip fix is preserved verbatim.
